+++ apps/demo/app/index.tsx
@@ -1,12 +1,11 @@
-import { useXUIColors, useXUITheme } from '@xaui/native/core'
-import { StyleSheet, View } from 'react-native'
+import { useXUIColors } from '@xaui/native/core'
+import { ScrollView } from 'react-native'
 import { useState, useEffect } from 'react'
 import { Button } from '@xaui/native/button'
 import { useRouter } from 'expo-router'
 
 export default function HomeScreen() {
   const colors = useXUIColors()
-  const theme = useXUITheme()
   const router = useRouter()
 
   const [, setValue] = useState(0.1)
@@ -19,14 +18,12 @@ export default function HomeScreen() {
   }, [])
 
   return (
-    <View
-      style={[
-        styles.container,
-        {
-          backgroundColor: colors.background,
-          gap: theme.spacing.md,
-        },
-      ]}
+    <ScrollView
+      style={{
+        paddingVertical: 32,
+        paddingHorizontal: 16,
+        backgroundColor: colors.background,
+      }}
     >
       <Button
         size="lg"
@@ -36,15 +33,6 @@ export default function HomeScreen() {
       >
         View Alert Examples
       </Button>
-    </View>
+    </ScrollView>
   )
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    alignItems: 'center',
-    justifyContent: 'center',
-    padding: 16,
-  },
-})