+++ apps/demo/CHANGELOG.md
@@ -1,5 +1,21 @@
 # demo
 
+## 1.0.20
+
+### Patch Changes
+
+- Updated dependencies [a4b7b3e]
+  - @xaui/native@0.0.9
+
+## 1.0.19
+
+### Patch Changes
+
+- Updated dependencies [976cea6]
+- Updated dependencies [7c8dde8]
+- Updated dependencies [dbbd7bc]
+  - @xaui/native@0.0.8
+
 ## 1.0.18
 
 ### Patch Changes